Conflicts:
	Makefile
	aom.cpp
	core.cpp
	mmath.cpp
	perfTest.cpp

+++ mmath.cpp
@@ -1,4 +1,5 @@
 #include "mmath.h"
+#include "mmath_const.h"
 #include <cmath>
 #include <numbers>
 #include <iostream>
@@ -18,7 +19,7 @@ namespace mmath{
 			double freq = startFreq + range * i / (double)size;
 			//std::cout<<"freq "<<freq<<std::endl;
 			for (int j=0;j<LENGTH_PER_FREQ;j++){
-				sin[i + j]=std::sin(2*std::numbers::pi_v<double>*freq*j/(double)SAMPLE_RATE);
+				sin[i + j]=std::sin(2*std::numbers::pi_v<float>*freq*j/(double)SAMPLE_RATE);
 				sin16[i + j]=MAX_VALUE*sin[i + j];
 				sin8[i + j]=((MAX_VALUE+1)/2-1)*sin[i + j];
